Extract degradation telemetry from the refresh module (#312)
Enrolling `scripts/` in the pylint gate put
`typos_rollout_refresh.py` over the 400-line module ceiling (438). The
bounded degradation counters and the HTTPS-only redirect handler form
a cohesive seam — telemetry and downgrade policy rather than cache
freshness — so they move to `typos_rollout_degradation.py` (120
lines), leaving the refresh module at 339.

Tests that exercise the counters or the redirect handler now consume a
`degradation_module` fixture; the freshness-validator tests keep the
`refresh_module` fixture.
